refactor dress code a bit
